Module Arbitrary__.Types_helper

val name : string -> string

Convention generator name for any type name

module Primitive : sig ... end

Module representing OCaml primitives types supported

val constr_type : loc:Ppxlib.location -> f:Ppxlib.expression -> args:Ppxlib.expression list -> unit -> Ppxlib.expression

constr_type loc f args () transforms an applicated parametrizable type into a QCheck arbitrary.

In that example, string is a parametrizable type applied on list

type t = string list [@@gen]

let gen_t = QCheck.list QCheck.string

val from_longident : loc:Ppxlib.location -> recursives_types:string list -> mutual_types:string list -> Ppxlib.longident -> Ppxlib.expression

from_longident loc recursives_types mutual_types lg transforms lg into a QCheck.arbitrary

Multiples cases:

The type is a identifier we use from_string

The type comes from an outside module, we require a generator inside that outside module

The type is an application, we raise an exception

val nest_gens : loc:Ppxlib.location -> Ppxlib.expression list -> Ppxlib.expression * string list * Ppxlib.pattern

Transform list of generators into a triple:

expressions nested with QCheck.pair expression

generators names used in the expression

pattern according to the nested expression

val record' : loc:Ppxlib.location -> gens:Ppxlib.expression list -> Ppxlib.label_declaration list -> Ppxlib.pattern * Ppxlib.expression * Ppxlib.expression

record' loc gens label_decls is an auxiliar function of record. It extracts the pattern and according expression for gens, also extracts the record's construction expression.

type t = left : int ; right = string

record' QCheck.int; QCheck.string left -> int ; right -> string =>

pattern : (arb_0, arb_1)

generators : pair QCheck.int QCheck.string

expression : left = arb_0 ; right = arb_1

val record : loc:Ppxlib.location -> gens:Ppxlib.expression list -> Ppxlib.label_declaration list -> Ppxlib.expression

record loc gens label_decls convert gens and label_decls in a application of a record type using QCheck.map.

Example:

type t = { left : int ; right : string } [@@gen]

let gen_t =
  QCheck.map (fun (x,y) -> { left = x ; right = y })
  (QCheck.pair QCheck.int QCheck.string)

val tuple' : loc:Ppxlib.location -> Ppxlib.expression list -> Ppxlib.pattern * Ppxlib.expression * Ppxlib.expression

tuple' loc arbs is a tuple auxiliar function, it extracts the pattern for a constructor application, the list of generators needed and the record expression

type t = int * string

tuple' int; string =>

pattern : (arb_0, arb_1)

generators : pair QCheck.int QCheck.string

expression : (arb_0 * arb_1)

val tuple : loc:Ppxlib.location -> Ppxlib.expression list -> Ppxlib.expression

tuple loc arbs converts arbs in a application of a tuple

Example:

type t = int * int * int

let gen_t =
  QCheck.map (fun (x,y,z) -> (x,y,z))
  (QCheck.triple int int int)

val constructors : loc:Ppxlib.location -> (Ppxlib.expression option * Ppxlib.expression) list -> Ppxlib.expression

constructors loc xs convert a list of (weight option * constructor) into a single expression choosing the constructor using it's weight (1 if it's not provided).

Example:

type t =
| A [@weight 5]
| B [@weight 6]
| C
[@@gen]

let gen_t =
  let open QCheck in
  frequency [ (5, always A) ;
              (6, always B) ;
              (1, always C) ]

val constructor : loc:Ppxlib.location -> kname:string -> ?⁠kargs:(Ppxlib.pattern * Ppxlib.expression * Ppxlib.expression) -> unit -> Ppxlib.expression

TODO: ocamldoc that comment

Convert a constructor name into an expression constructor QCheck.arbitrary

Example:

type t =
| A

(* A => QCheck.make @@ QCheck.Gen.return A *)

An additional case is supported when constructor requires arguments

Example:

type t =
| A of int * string
| B of { left : int ; right : string }

(* A => QCheck.map (fun (x,y) -> A (x,y)) QCheck.(pair int string)
   B => QCheck.map (fun (x,y) -> B { left = x ; right ; y }) QCheck.(pair int string) *)

val tree' : loc:Ppxlib.location -> leaves:Ppxlib.expression -> nodes:Ppxlib.expression -> unit -> Ppxlib.expression

tree' loc leaves nodes () is almost the same function as tree' the only difference is that QCheck.frequency is already applied to leaves and nodes

val tree : loc:Ppxlib.location -> leaves:(Ppxlib.expression option * Ppxlib.expression) list -> nodes:(Ppxlib.expression option * Ppxlib.expression) list -> unit -> Ppxlib.expression

tree loc leaves nodes () transforms a tree type like into a recursive arbitrary expression

The recursive arbitrary uses a fuel, we could imagine that in future work the fuel would be provided by the user.

Example:

type t = Tree | Node of int * leaf * leaf
[@@deriving arb]

let rec arb_tree fuel =
  let open QCheck in
  match fuel with
  | 0 -> frequency [(1, always Leaf)]
  | n ->
    frequency
      [
        (1, always Leaf) ;
        (1, map
          (fun (arb_0, (arb_1, arb_2)) -> Node (arb_0, arb_1, arb_2))
          (pair int
             (pair (arb_tree (n - 1)) (arb_tree (n - 1)))))
      ]

let arb_tree = arb_tree 5

val variants : loc:Ppxlib.location -> ty:string -> [< `RTag of string * Ppxlib.expression option * Ppxlib.expression list | `RInh of Ppxlib.expression option * Ppxlib.expression ] list -> Ppxlib.expression

variants loc ty xs create a QCheck.arbitrary using xs to produce Ptyp_variant _ list. We also require ty to constraint the expression to a

t QCheck.arbitrary 

where

t 

is the current type we are deriving.

`RTag represents the direct declaration of a variant

type t = [`A]

`RInh represents an inheritage of another variant

type t' = [`B | t]

Each variant can have a specific weight, see constructors.

val gen : loc:Ppxlib.location -> is_rec:bool -> args:Ppxlib.pattern list -> ty:string -> Ppxlib.expression -> Ppxlib.structure_item

gen loc is_rec args ty body create a QCheck.arbitrary for ty

When is_rec is false we produce the following structure_item:

let arb_ty args = body 

The specific case where is_rec is true, the generator is self_recursive and we produce the following expression

let arb_ty () = arb_ty' 5
and arb_ty' fuel =
  match fuel with
  | 0 -> (* leaves *)
  | n -> (* recursives calls *)
and arb_ty = arb_ty ()

(TODO: I did not test self recursive type with parameters inside args)

(TODO: name gen is misleading, for now we produce an arbitrary but this will be later on transformed into a Gen.t)

val gens : loc:Ppxlib.location -> tys:string list -> gens:Ppxlib.structure_item list -> unit -> Ppxlib.structure_item

val observable : loc:Ppxlib.location -> Ppxlib.core_type -> Ppxlib.expression

observable loc x create an QCheck.Observable.t if x is an observable

val fun_nary : loc:Ppxlib.location -> Ppxlib.expression list -> Ppxlib.expression -> Ppxlib.expression

fun_nary loc obs x create an ('a Tuple.t -> 'b) fun_ arbitrary