Module Arbitrary__Attributes

val arb : Ppxlib.core_type -> Ppxlib.expression option

arb loc ct look for an attribute "arb" in ct

example:

type t =
| A of int
| B of (int [@arb QCheck.int32])

It allows the user to specify which arbitrary he wants for a specific type. Returns the arbitrary as an expression and returns None if no attribute is present

val weight : Ppxlib.attributes -> Ppxlib.expression option

weight loc ct look for an attribute "weight" in ct

example:

type t =
| A [@weight 5]
| B [@weight 6]
| C

It allows the user to specify the weight of a type case.