Module Arbitrary__Recursives_helpers

type ty = string

val is_recursive_row_field : loc:Ppxlib.location -> ty -> Ppxlib.row_field -> bool

is_recursiverow_field ty rw traverses rw looking for ty

val is_recursive_row_fields : loc:Ppxlib.location -> ty -> Ppxlib.row_field list -> bool

is_recursiverow_fields ty rws traverses rws looking for ty

val is_recursive_constructor_declaration : loc:Ppxlib.location -> ty -> Ppxlib.constructor_declaration -> bool

is_recursive_constructor_declaration ty constr_decl traverses constr_decl looking for ty

val is_recursive_type_declaration : ?⁠loc:Ppxlib.location -> Ppxlib.type_declaration -> bool

is_recursive_type_declaration td does a tree traversal of td and returns true if the type inside td is self recursive.

val get_recursives_type_declarations : ?⁠loc:Ppxlib.location -> Ppxlib.type_declaration list -> ty list

get_recursives_types_declarations tds returns the sub-list of type_declaration inside tds where is_recursive_type_declaration returns true