Module Common__Ast_helpers.Pattern

val pattern : ?⁠loc:Ppxlib.location -> ?⁠loc_stack:Ppxlib.location_stack -> ?⁠attributes:Ppxlib.attributes -> Ppxlib.pattern_desc -> Ppxlib.pattern

val ppat_var : ?⁠loc:Ppxlib.location -> ?⁠loc_stack:Ppxlib.location_stack -> ?⁠attributes:Ppxlib.attributes -> string -> Ppxlib.pattern

Build a Ppat_var

val ppat_any : ?⁠loc:Ppxlib.location -> unit -> Ppxlib.pattern

Build a Ppat_any

val extract_pat_name_opt : Ppxlib.pattern -> string option

Extract pattern name, returns None if the name can not be found

val extract_pat_name_exn : loc:Ppxlib.location -> Ppxlib.pattern -> string

Extract pattern name, raises an exception if the name can not be found