Module Ppx_deriving_qcheck

val derive_gens : version:[ `QCheck | `QCheck2 ] -> loc:Ppxlib.location -> (Ppxlib.rec_flag * Ppxlib.type_declaration list) -> Ppxlib.structure

derive_gens ~version ~loc xs creates generators for type declaration in xs.

The generators can either be QCheck.Gen.t or QCheck2.Gen.t based on version.

val derive_arbs : loc:Ppxlib.location -> (Ppxlib.rec_flag * Ppxlib.type_declaration list) -> Ppxlib.structure

derive_arbs ~loc xs creates generators for type declaration in xs and use these lasts to build QCheck.arbitrary.